requirements.txt
alembic==1.9.1
aniso8601==9.0.1
anyio==3.6.2
asyncpg==0.27.0
attrs==22.2.0
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.88.0
graphene==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.1
packaging==22.0
pluggy==1.0.0
pydantic==1.10.4
pytest==7.2.0
pytest-asyncio==0.20.3
python-dotenv==0.21.0
python-gql==0.2.4
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a31
sqlmodel==0.0.8
starlette==0.22.0
starlette-graphene3==0.6.0
starlette-graphql==0.2.1
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
